+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+## Added
+
+- Command to open the PreTeXt getting started walkthrough.
+
+## Fixed
+
+- Fixed bug that prevented `installSage.sh` from running in a codespace.
+
 ## [0.27.0] - 2025-04-07
 
 ### Added